SpotAnalyzerFactory can be told the image they will work on.
This method is only used by some factories that create feature keys
depending on the source image. For instance if an analyzer must declare
one feature value per channel, it needs to know how many channels there
is in the source image. This method is here to pass this information.

The information is passed on by the SpotAnalyzer provider.
This change breaks the API, because now the SpotAnalyzerProvider must
be created with a reference to the source ImagePlus (null is ok though).