Spot feature analyzer for multi-channel mean intensity measurements.
In Benoit Lombardo initial attempt:

 https://github.com/tinevez/TrackMate-extras/blob/master/src/main/java/fiji/plugin/trackmate/extra/spotanalyzer/SpotMultiChannelIntensityAnalyzer.java

We could not know the number of channels so we generated 10 feature
at once. This version uses the fact that now we can now the number of
channel we work on. Also, since it is now an IndependentSpotFeatureAnalyzer
it only recomputes the spot features it needs to.